tests profiles and SonarCloud setup

+++ src/main/java/net/engineeringdigest/journalApp/service/Impl/JournalEntryServiceImpl.java
@@ -1,11 +1,14 @@
 package net.engineeringdigest.journalApp.service.Impl;
 
+import lombok.extern.slf4j.Slf4j;
 import net.engineeringdigest.journalApp.entity.JournalEntry;
 import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import net.engineeringdigest.journalApp.service.JournalEntryService;
 import net.engineeringdigest.journalApp.service.UserService;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,6 +18,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class JournalEntryServiceImpl implements JournalEntryService {
     @Autowired
     private JournalEntryRepository journalEntryRepository;
@@ -45,11 +49,22 @@ public class JournalEntryServiceImpl implements JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(id);
+
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch(Exception e) {
+            log.error("Encountered this error : ", e);
+            throw new RuntimeException("An error occurred while deleting the entry : ", e);
+        }
+        return removed;
     }
     public JournalEntry updateById(ObjectId id, JournalEntry journalEntry) {
         journalEntryRepository.save(journalEntry);
